backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-core>=0.1.10
langchain-ollama>=0.0.1
graphviz>=0.20.1
requests>=2.31.0
python-dotenv>=1.0.0
ollama>=0.1.5
pydantic>=2.5.2
python-multipart>=0.0.6
websockets>=12.0
pillow>=10.0.0
